requirements.txt
matplotlib
pandas
Pillow
streamlit
openpyxl
streamlit-authenticator
plotly
numpy
emoji
pymysql
